+++ pdi/src/expression/impl/operation.h
@@ -1,5 +1,6 @@
 /*******************************************************************************
  * Copyright (C) 2015-2020 Commissariat a l'energie atomique et aux energies alternatives (CEA)
+ * Copyright (C) 2020 Institute of Bioorganic Chemistry Polish Academy of Science (PSNC)
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -53,11 +54,13 @@ struct PDI_NO_EXPORT Expression::Impl::Operation: Expression::Impl {
 	
 	using Operand = std::pair<Operator, Expression>;
 	
-	
 	Expression m_first_operand;
 	
 	std::vector<Operand> m_operands;
 	
+	Operation();
+	
+	Operation(Expression first_operand, Operator op, Expression secend_operand);
 	
 	std::unique_ptr<Impl> clone() const override;