+++ pdi/CHANGELOG.md
@@ -5,17 +5,34 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### For PDI user
 
-### Added
+#### Added
 
-### Changed
+#### Changed
 * Require the CONTIGUOUS attribute for all Fortran arrays (including 1D)
 
-### Deprecated
+#### Deprecated
 
-### Removed
+#### Removed
 
-### Fixed
+#### Fixed
 * Added the CONTIGUOUS attribute to all Fortran arrays in tests
 
-### Security
+#### Security
+
+
+### For plugin developer
+
+#### Added
+* Added expression arithmetic operators
+
+#### Changed
+
+#### Deprecated
+
+#### Removed
+
+#### Fixed
+
+#### Security